Working #24: Error handling now works.
Error handling now works on the front index page. Hoping to find a way to refine it using Multer but it is functional for now. Upload is inturrupted if user does not upload two files or if they upload an invalid file type (restricted to MEI and PNG for now.