Reworking revert
Had to change the file structure for revert so that the original MEI file was not lost everytime you left the editor page. The code still needs to be optimized to remove reused code. Possible solutions are to use regex expressions or to create a seperate function to handle actions with all three folders.