Refine when we show grouping options.
Don't give the option to group when the selection is already grouped.
Also give the next-highest grouping option if the selection makes up
that entire higher-level group. For example if selecting by NCs and an
entire neume is selected, give the neume option. If selecting by NCs and
part of a neume is selected (but all the NCs are in the same neume),
give no options.

Resolves #251.