Support two-finger panning with a touchpad
Actually the touch actions are only for touch screens, not touch pads.
So we have to use the wheel action instead. This will mean a user
scrolling with an actual wheel will still be able to pan. For some
reason the scroll action does not occur on the right element, only
wheel.

Resolve #261.