Fix select/drag issues from refactor/select merge.
Call `dragHandler.dragInit` on every new select action. Do not click
reselect on a part of an existing selection (since this is probably
dragging a drag selection, not making a new selection). Fix an issue
where we called a jquery method on an HTMLElement.

Check manually before merging into develop for correctness. It looked
good on my end but I might've missed something.