Support dragging staves from any part of staff
Now dragging a staff and its elements can be done by clicking anywhere
within the staff and not just on the actual SVG elements. This was done
by sending a `mousedown` action to the staff SVG when click selecting
off the SVG and, in that case, temporarily setting the dragging actions
to use `#svg_group` as a source for `mousemove` and `mouseup` actions.

Resolve #296.