Support zooming by scrolling while holding Shift
Adds a `scrollZoom` function in the ZoomHandler that does this, and will
trigger dragging if it's actually the regular scrolling/touchpad
panning. Gets min and max values for zoom directly from the slider DOM
element's min and max values, so it should update to any new bounds.
Values for the slider updated automatically.

Resolve #300.