Switch to using PouchDB, WIP
Since the local storage is either 5 MB or 10 MB depending on browser, we
have to use indexedDB to store the background image. That is hard, so
we're using PouchDB instead.

Currently we are able to get the MEI from the database, but as of now
the background image isn't being displayed (although the data appears to
be stored).

Note that right now we are only able to write to the database once,
since updating PouchDB is complicated and requires references to the
revision.