Disable loading animation at part of execution.
If we support Safari we can't rely on the load event to be triggered for
SVG elements. Even though this does seem to be the behavior defined by
W3, it just doesn't work in Safari (and presumably other webkit
browsers).

Code for the animation is kept in case we can find a more permanent
solution, but since we want to move to diva.js and won't be directly
loading images this isn't a high priority.

Resolve #320