Encode MEI map so we can pass it using express.
Express can only send parameters as a string. The way to do this is to
convert the map generated on the server end into an array and then
encode that array as a URI component so it is not altered while sending.
Then decode the URI component and parse it as an Array object. Then use
the Array as the parameter in the Map constructor.