Write custos outside of sb
From #604 (also see music-encoding/music-encoding#629), the guidelines
allowing a custos in a sb are actually in error. These changes still
allow loading files with a custos in a sb (as those WERE considered
correct by the schema) but will write the custos as a child of the
layer/syllable as the sibling preceding the sb.