Remove update-verovio script
This script is outdated considering the many changes to the structure of
Neon since its writing. Since we're also planning to switch to only
using upstream releases of Verovio after the first release incorporating
our editor changes, there isn't much of a point in maintaining this
file. Likewise keeping it here is confusing.