🐛 Fix @follows/@precedes check with Array.find()
Previously, we were finding the next syllable with positional indices (hence, n+1th syllable should be the following syllable for the nth syllable), which doesn't work when there are syllables in between the toggle-linked syllables.

The solution is to find the syllable in the array by the ID instead, which handles this edge case.

The time complexity stays the same, linear time.