🔨 Set diva.js to a specific release
diva.js being linked to the GitHub link kept giving me errors about not having my SSH key authenticated.

Juliette has said that it was linked to the latest version of diva.js because they were making changes to diva.js at the time.

Since we are not making changes to diva.js currently, I think it would be better to settle on the latest release.