🐛 Fix non-existent updateSelectedBBox()
There was an error with updateSelectedBBox() which would be called any time the user edits the syl text of a bounding box. This issue seemed to have been due to some misorganization and perhaps discarded during a hasty merge of a pull request.

Anyways, it has been extracted into its own method which can be called from both TextEditMode.ts and ModalWindow.ts