♻️ Restructure drag tests
In regards to the reselect-dragged-fix, I added checks on whether the element is reselected after it has been dragged.

One other thing that has been changed in this commit is:

Drag tests have been crashing, and I think it's due to the sheer number of them in one file. I also set `numTestsKeptInMemory` to 0, which may help with this crashing issue.

Related: https://github.com/cypress-io/cypress/issues/1906